Break up the tests a bit
