+++ uobtheatre/graphql_auth/test/test_decorator_helpers.py
@@ -1,0 +1,6 @@
+from uobtheatre.graphql_auth.test import decorators as test_decorators
+
+
+def test_graphql_auth_test_decorator_helper_returns_skipif_mark():
+    mark = test_decorators.skipif_django_21()
+    assert mark.mark.name == "skipif"
